chore: consolidate type hierarchies, fix BookStatus enum (closes #133)
- Fix BookStatus: 'want_to_read'/'currently_reading'/'read' → 'TO_READ'/'READING'/'READ' (match backend)
- Add types/admin.ts and types/auth.ts; barrel-export from types/index.ts
- Migrate 22 files off @/interfaces/ → @/types/
- Add apiFetch/apiPost/apiPut/apiDelete wrappers to lib/api-client.ts
- Migrate 8 files off @/utils/api → @/lib/api-client
- Delete src/interfaces/, src/utils/api.ts, src/utils/auth.ts, src/lib/book-utils.ts
