--- frontend/src/interfaces/user.ts
@@ -1,7 +1,0 @@
-export default interface User {
-  id: number;
-  name: string;
-  email: string;
-  profile_picture?: string;
-  is_admin: boolean;
-}
