+++ frontend/src/lib/index.ts
@@ -1,6 +1,5 @@
 // Barrel exports for lib utilities
 export * from './api-client';
 export * from './auth';
-export * from './book-utils';
 export * from './constants';
 export * from './utils';
