+++ frontend/src/types/index.ts
@@ -1,5 +1,7 @@
 // Barrel exports for all type definitions
+export * from './admin';
 export * from './api';
+export * from './auth';
 export * from './book';
 export * from './review';
 export * from './user';
